+++ internal/query/client.go
@@ -174,7 +174,10 @@ func (c *Client) ExecuteScript(
 		),
 	}
 
-	request, grpcOpts := executeQueryScriptRequest(a, q, settings)
+	request, grpcOpts, err := executeQueryScriptRequest(a, q, settings)
+	if err != nil {
+		return op, xerrors.WithStackTrace(err)
+	}
 
 	op, err = executeScript(ctx, c.client, request, grpcOpts...)
 	if err != nil {